+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 xxxxxxxxxxxxxxxxxxx
+===================
 - Added `Apify.amIAtHome()` function that returns `true` when code is running at Apify platform and `false` otherwise
   (for example locally).
+- Added `ignoreMainProcess` parameter to `AutoscaledPool`. Check documentation for more info.
 
 0.5.26 / 2018-04-27
 ===================